refactor(editor): workspace root 분리
